requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.2.1
flasgger==0.9.7.1
Flask==3.1.1
freezegun==1.5.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
six==1.17.0
typing_extensions==4.14.1
Werkzeug==3.1.3
